[FIX] web_cohort: exclude empty data from average computation
The average was computed considering null values as zeros.
Say a ticket that has a rating of 5 and another one that has no rating,
the resulting avergae would be (5 + 0) / 2 = 2.5.
With this commit, we don't take into account null values.
In other words, we divide by 1, so the average is 5.

Also, for fields with group_operator="avg", we had a _read_group that
would get the average to put it in the measure column. The issue was
that we used the same value to compute percentages i/o the sum.
In retention mode, we start from 100% and subtract the values.
Say 4 tasks are created on d-day, 3 are done on d+1, 1 on d+2.
We have measure that represents how many tasks are left on each day.
Previously, we would calculated the average (3 + 1) / 2 = 2,
and from that, calculate the percentage on d+1 as (2 - 3) / 2 = -50%.
With this commit, we use the sum (4) to do it: (4 - 3) / 4 = 25%.

task-3475463

Part-of: odoo/enterprise#46693