[IMP] hr_expense_extract: usability improvements
The method _autosend_for_digitization was called twice when attaching a receipt in an expense. Once due to the widget `attach_document` and another due to an action with the same name. This commit moves the action to the community repo and changes it so it sets the main attachment for the expense instead.

task-3539382

closes odoo/enterprise#50881

X-original-commit: d8bbbcb0d01aae063272fea5531bdfcfc4105f8e
Related: odoo/odoo#142451
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Ali Alfie (alal) <alal@odoo.com>