[IMP] account_reports: composite reports
The goal of this PR is to add a button on the reports next to the filters that
will redirect to the configuration page of the report.
On this page we have changed a placeholder and hide the use_section field in
case the report has lines.

closes odoo/enterprise#50235

Task: 3569038
Related: odoo/odoo#141103
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>