[FIX] helpdesk_timesheet: fix default company_id for timesheet
When creating a timesheet from the ticket view form, an access error is
raised if the active company is different from the company of its
project.

Step to reproduce:
- install helpdesk_sale_timesheet
- open helpdek
- open configuration menu -> helpdesk teams
- select a team with the visibility set on 'invited portal users'
- enable the 'billing' feature, and add a project on this team
- select as active company any company other than the one set on the
project selected before
- select an existing ticket, or create a new one on the team
- try creating a new timesheet
Validation error is raised

Source of the issue:
If no company_id is given in the vals_list, the default company_id used
by analytic line model is the active company. The issue is not present
on task, because the key 'company_id' is correctly set in the
'_timesheet_preprocess' method. This method is overwritten in the
helpdesk_sale module and the updates done in it prevent the correct
setting of the 'company_id' later on.

Solution:
Ensure that the 'company_id' key is present and consistent when creating
a new timesheet from the ticket form view.

version: 16.0 - master

task - 3581941

closes odoo/enterprise#65575

X-original-commit: 888aa4c57b26b8d3fc4c2e421a385a8c31044d16
Signed-off-by: Damien Raquet (damr) <damr@odoo.com>
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>