[FIX] planning: avoid traceback when displaying employee avatar
On some db without demo data and in particular configuration, the
changes introduced in https://github.com/odoo/enterprise/pull/49333
create a problematic situation where no progress bar info is available
for a row with an employee.

This commit avoids such traceback by falling back on the relation used
for employee_id field if no info is available in the progress bar on the
model used.

closes odoo/enterprise#50396

X-original-commit: bc9a73ff23143de72384b4351ea72bcb0f19d0d8
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>