[FIX] web_studio: prevent traceback when clicking on the reified field
Currently, an error occurs when the user tries to select a
selection, boolean, or radio field.

Steps to produce an error:
1. Install web_studio
2. Go to Settings > Users & Companies > Users > Select any user
3. Open studio mode by clicking the studio icon.
4. Try to select the Administration field from the access right page

This is because before the owl refactor of webstudio with Commit[1],
the code tries to access the field name from the 'attrs' of the node (see [1]),
but after Commit[1], it tries to access the field name from the 'field' of the
node (see [1]), and it works for normal fields like text, number, etc., but not
for grouped fields like selection, boolean, radio, etc.

This commit fixes the above issue by not fetching or changing the
default value of reified fields. Now, after this commit, the default
value box is not showing for the reified fields.

Commit[1]-https://github.com/odoo/enterprise/pull/37285/commits/e44403193cbfd7893708ff739fd072f8bb1197c6
[1]-https://github.com/odoo/enterprise/blob/1fd7b1182146f1d7e6a7cbb9b725cb4d47281e14/web_studio/static/src/legacy/js/views/view_editor_manager.js#L1029
[2]-https://github.com/odoo/enterprise/blob/50b6e04d2e2f47797ec14bf506a030bd1d9bc797/web_studio/static/src/client_action/view_editor/interactive_editor/properties/field_properties/field_properties.js#L117

opw-3924602
sentry-4156923959

closes odoo/enterprise#63729

X-original-commit: eaf1a33c50a84602063fadf8bb1ffbc77a8c61fd
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Hubert Van De Walle <huvw@odoo.com>
Co-authored-by: Hubert Van De Walle <huvw@odoo.com>