[FIX] sale_subscription: archived pricelists from the products
Steps to produce:
- create a new price list
- add the product to the price list
- archive the price list
- go to that product
- the archived price list is available

The product will be removed from the pricelist when archived,
but it won't be removed from the product when archived.
This commit has fixed the issue. Once the pricelist is achieved,
it will be removed from the product and once the pricelist
is unachieved, it will be added to the product.

task-3573491

closes odoo/enterprise#50262

Related: odoo/upgrade#5451
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>