[FIX] frontdesk: spacing in form view
The o_field_image in the Drinks in the form view is cancelling the
negative margin applied by the row class on the o_group.

Using a gutter-x 0 on the o_group removes the padding allowing the
inputs to be aligned with the title. This is a quickfix and the behavior
of the o_field_image in a form view should be reviewed consistently
across Odoo because different module seems to use their own fix for
this.

task-3584529

closes odoo/enterprise#50265

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>