[IMP] account_accountant: allow unknown fields in chart template.
Problem
---------
In most cases, default deferred accounts and journal need to be set up
for localizations. This is normally done in the enterprise report module
for that localization. However, in some cases, the localization does not
have special report formats. In such situation, a localization report
module that sets up very few default values for the data company is
defined. This is way overkill.

Objective
---------
Allow the community company template to have 'unknown fields' defined.
Doing so, allows for the default deferred accounts and journal to be
defined without entreprise module to exists. Currently, this raises an
error.

Solution
---------
Override the `_get_chart_template_data` in `account_accountant` so that
it fetches the values for deferred accounts/journal that may be defined
in the community company templates. It uses the values defined in the
templates by default. If not found, it fetches for an relevant account/
journal for deferred moves in the database in the rest of the template
definition.

closes odoo/enterprise#50305

Related: odoo/odoo#138937
Signed-off-by: William André (wan) <wan@odoo.com>