[IMP] account_asset, account_asset_fleet: vehicle
Before this commit, it was not possible to add a vehicle to a newly created asset. Now in debug mode, you can find the vehicle in the value import group of the form view.

Two flows are possible, either creating a move with a vehicle and creating an asset from it. There the vehicle in the asset page is fill with the value of put in the move. Or creating directly an asset from scratch and putting a vehicle (there the associate journal entry must have the right vehicle).

closes odoo/enterprise#50279

Task: 3581933
Signed-off-by: William André (wan) <wan@odoo.com>