[IMP] account_reports: placeholder on drag
With this commit, when dragging an element on the creation/edition of a
report, it will keep a placeholder at the original place. To do that, we
keep a copy of the element that is being dragged. The copied element is
sanitized which means that we remove all attributes that could create a
conflict with any other element.

Task-id 3582082

closes odoo/enterprise#50283

Related: odoo/odoo#145027
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>