[FIX] industry_fsm : default planned dates
Before this commit, the default planned dates were not set when creating
a task from the gantt view in field service.

As the default planned dates when creating a task from fsm should not
be different than from project, we delete the default_get code
concerning the planned dates in FSM and we apply the same default as for
normal tasks.

task-3576701
task-3586078

closes odoo/enterprise#50448

X-original-commit: 713486dcccbaf97a46ff03a4ef1569d335e5d762
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>