[FIX] account_reports: foreign currency symbol on partner ledger report
Issue:
======
In `Parnter Ledger` report it only shows the primary currency symbols and not the foreign currency (If the bill was in foreign currency)

Steps to reproduce the issue:
=============================
- Make vendor bill in `Accounting` app in a different currency than the journal
- Open `Parnter Ledger` report (`Accounting` app > Reporting > Partner Ledger
- The `Amount Currency` field will be debited in the journal currency, not the one you specified. (just a wrong symbol, digits are fine).
-
This was due to the transition from version 16.3 > 16.4, in which currency symbols started to appear in the report (in previous versions symbols appeared only if it was a foreign currency), but the logic in the partner ledger remained the same, while the principle of passing a string to the report itself changed.

Solution
========
I changed the logic of passing currency in report._build_column_dict by adding currency passing to the function and removing variables that are not used.

opw-3584282

closes odoo/enterprise#51595

X-original-commit: d4562f2904373bb85893d1ae5708010c90a4b4b9
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Ilya Rudy (ilru) <ilru@odoo.com>