[IMP] documents: improve document preview layout
Improve the document preview layout by setting the inspector buttons
above the document preview to separate them from the document fields.

Move the trash button next to the document name for understanding purpose.
When a document is archived the restore and trash buttons are displayed
above the preview to leave the document name some space.

Task-3566683

closes odoo/enterprise#49537

Signed-off-by: Stéphane Debauche (std) <std@odoo.com>