[FIX] sale_planning: avoid opening plan dialog if no shift to plan
Before this commit, when the user clicks on a gantt cell to plan a
shift, a plan dialog is displayed even if there is nothing to plan.
That behavior could be annoying for the user if he just wants to
plan a new shift each time he clicks on a gantt cell.

This commit checks if there is at least one shift to plan, if yes
then the plan dialog will be displayed otherwise the form view dialog
will be displayed to create a new shift to plan.

closes odoo/enterprise#50468

X-original-commit: eb2eb0bb951225b1c488f8580707d8c135cbfce0
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>