[FIX] account_reports : general ledger localization
Issue:
======
In some countries (like Italy), with the corresponding l10n module installed,
opening the General Ledger and selecting the "Customer Invoices" or "Vendor Bills"
journal in the filter crashes, telling the report_id of the computed options
is inconsistent with the report using them.

Steps to reproduce the issue:
=============================
- Install accounting and l10n_it_reports
- Go to accounting / reporting / general ledger
- Choose customer invoices in journals

Origin of the issue:
====================
When selecting only sales or purchase journals (not both together),
the General Ledger displays additional lines corresponding to the data
of the Generic Tax Report matching the chosen options. The way it was
done in 16.4, however, let the tax report choose the first variant
available for the current company, so if the l10n module contained a
tax report, the wrong report was chosen and set in report_id, and we
were calling the Generic Tax Report with an options dict whose
'report_id' key corresponded to a localized variant instead of the root
report.

Solution:
=========
Previous versions used to directly force the 'report_id' key, but this
does not work anymore because of the introduction of the composite
reports (aka sections), and the way they changed the options generation.
The key to force is now 'selected_variant_id', to make sure the root
report is the one that gets selected.

opw-3580078

closes odoo/enterprise#50554

X-original-commit: 19af3fd86e10edf4f3f131289ae8d8c4cd68041a
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Mahdi Cheikh Rouhou (macr) <macr@odoo.com>