[FIX] web_studio: non-existing method called
The customer wants to use Web Studio to create an add custom button from an
existing method in the model; however, there is a non-existent method identified,
at that time the error will be generated.

step to reproduce:
- install the `web_studio & sale_management`
- open sale
- open any one Quotations
- click on `web_studio` button
- click on Add button
- enter the label & select `call a method`
-  write method name `action_delete` (non-existing method in model)
- the error will be generated,

after this commit ,If the customer prefers a non-existing method,at that time
raise Validation Error.

sentry-4614276017

closes odoo/enterprise#50890

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>