[FIX] documents: inspector change are not saved when record is not selected
Bug
===
When we open the document inspector without selecting the record,
if we change some field values, those changes are not saved.

Technical
=========
Since 680d61b9ee9cfbeec679dd07e1cfe48ea60162e5 , we do not select the
records when we preview them. The purpose of that was to remove the code
that save the selection to restore it when we close the preview, etc.

However, the web client won't save automatically the records if they
are not selected (see relational_model.js@_update).

So to fix that issue, we save ourselves in Documents when we make
changes on a record which is not selected.

Task-3586696

closes odoo/enterprise#51877

X-original-commit: 19711bb280317b55044203e0de118c46febaed6f
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>