[FIX] mail_mobile,*: open "direct message" on Mobile doesn't work
* = test_mail_enterprise

Since commit [1], the model `mail.channel` was renamed to
`discuss.channel`. In our Mobile App (Android, iOS) we have a specific
logic to handle opening a DM with the correct action related to this
model. Due to this renaming, we were not able to open the action of the
direct message on Android and iOS apps.
Also, as we can't push change on iOS, we are not able to adapt the
logic in the Mobile App (due to Apple Store Policies).

This commit forces to send the old model to apply the correct logic
on the Mobile App (Android, iOS).

Steps to reproduce:
* Send a DM to the Mobile App's user
* The user clicks on the notification
* The Mobile App opens but not in the correct view => BUG

[1]: odoo/odoo@90cb44e1e1b6bad9c12f3bd14d1c9e340461e19e

closes odoo/enterprise#50586

X-original-commit: 127cf0f5e8f63408b2116d1b543717fac8971973
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>