[FIX] marketing_automation: avoid trace duplicates after campaign copy
Copying the 'last_sync_date' field when copying a campaign resulted in the
copied activities to be considered new activities. When an update is performed
on such a campaign this results in traces being duplicated for all activities
present when the campaign performed its first participant synchronization.

Simply setting the last_sync_date as a non-copied field solves the issue.

Task-2981581 (MA: Fix trace duplication and various issues)
opw-2643368

X-original-commit: d22ad9f4392e93373b9d828783ef1a61647c79b1
Part-of: odoo/enterprise#51838
Co-authored-by: Stéphane Debauche <std@odoo.com>
Co-authored-by: Thibault Delavallée <tde@odoo.com>