[IMP] account_saft_import,l10n_{dk,lt,ro}_saft_import: add modules
Allow a user to import his data with SAF-T.
We allow import of taxes/partners/accounts/journals/moves.
Added specifities for DK, LT and RO in specific modules

We consider:
- a partner to be unique per name + VAT number.
- an account to be unique with his code.
- a tax to per name + amount_type + amount + country_id

The journal and moves are always created the first time, and then, we use the ID defined in the SAF-T to avoid recreating them (especially important for journals).

We create opening balances for accounts if the user checked the option.

task-3593719

closes odoo/enterprise#51661

Related: odoo/odoo#143986
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>