[FIX] l10n_mx_edi,whatsapp: speed up ir.attachment deletion
Two many2one to ir.attachment were added on table that can have
a lot of record in there respective table.

The deletion of ir_attachment need to check the new foreign constraints
on those table which can take a lot of time.

It can block the autovacuum since _gc_lost_attachments can timeout

To speed up the delete, btree indexes neet to be added on many2one
that point to ir_attachment

closes odoo/enterprise#50714

X-original-commit: 63d3ec65594f51f2396845b722b370d8820c3ba0
Signed-off-by: Jérémy Kersten <jke@odoo.com>
Signed-off-by: Thibault Francois <tfr@odoo.com>