[FIX] social_linkedin: improve the preview of the article share
Purpose
=======
When we share an article on LinkedIn, a preview of the page is
displayed. If we post the same article from Odoo, the preview is not
displayed, which has an impact on how much engagement the post will get.

We also show the thumbnail of the URL in the social post form view and
we download the thumbnail image to show the preview in the feed view.

Technical
=========
To build the preview, we make an HTTP request to the URL, and extract
some metadata of the page in the <meta/> tags). We have to upload the
image to the LinkedIn digital assets in order to use that image as a
thumbnail when we create the post.

Task-3594550

closes odoo/enterprise#52458

X-original-commit: f24a0487fef31fba72923fe4e07fde2a1fad28bb
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>