[FIX] marketing_automation: avoid trace duplicates for new participants
When creating a new participant in a running campaign, traces are created
automatically for all 'begin' activities. This includes new activities though
which don't have any traces yet associated with existing participants.

As a result, duplicate traces would be introduced on the next activity
synchronization (action_update_participants) because of this inconsistency
between participants that existed before the activity creation and participants
that did not.

This commit introduces a check so that the activity synchronization now checks
for existing traces before creating new ones.

Task-2981581 (MA: Fix trace duplication and various issues)
opw-2643368

X-original-commit: e76bc99ec094ca30ed74a5cc5dd21182e61d2cf7
Part-of: odoo/enterprise#51838
Co-authored-by: Stéphane Debauche <std@odoo.com>
Co-authored-by: Thibault Delavallée <tde@odoo.com>