[FIX] l10n_fr_fec_import: performance
Do not recompute the whole CoA data in a loop. On top of building huge
mappings and lists and everything, it also does database operations.
Extracting the building of the data out of the loops makes an import of
80k lines from 15 minutes to 1 second.

Also, avoid trying to create partners without name. If there is a
reference in the file (`CompAuxNum`), we can use that when the label is
empty (`CompAuxLib`)

X-original-commit: d151abe782b917d227a833f9b5c01f645dd3a071
Part-of: odoo/enterprise#51027