[FIX] stock_barcode: barcode rule conversion
Before this commit, the rules from the default barcode nomenclature
wasn't used in the Barcode App, except for the weight, and it wasn't
considered as a big deal, but since user can receive an EAN-13 barcode
for a product with an UPC-A barcode in the DB, we can use the rule to
ensure the conversion since barcode nomenclatures already have a
parameter, `upc_ean_conv`, for this purpose.

How to reproduce:
- Create a product using an UPC-A as barcode (eg.: 123123123125);
- Open the default barcode nomenclature and ensure the UPC/EAN
  Conversion is set on "Always" or "EAN-13 to UPC-A";
- For this nomenclature, add a rule to catch an UPC-A (eg.: UPC-A as
  encoding and "\d\d\d\d\d\d\d\d\d\d\d\d" as the pattern). This rule is
important because the conversion only happens when the converted barcode
match a rule. Also, ensure this rule will not be masked by any other
rule (for example, the default one for the product barcodes since this
rule uses a pattern to match any barcode);
- Go in the Barcode App, create a new receipt and scan the product's
  barcode but as a EAN-13 (just add a 0 before, eg.: 0123123123125)
  -> The product is not found.

To fix this issue, if the parser returns a barcode of the "Unit Product"
type but different than the scanned one, the Barcode App will use this
one instead of the scanned one. That enables the EAN-13 to UPC-A
conversion, but the kiss cool effect is that also allows to use the
alias in the Barcode App :)

task-3472884

X-original-commit: 4b32e7c8c1417d577a47e7350bf15e09d9dbff2e
Part-of: odoo/enterprise#53600