[IMP] account_reports: Handle account tags in Coverage tool
The aim of this commit is handling account tags used in
Balance Sheet/Profit and Loss report on the Coverage report
tool.
The idea is that user can change how we retrieve the account
for a specific expression, the commit handle the case where
we change for a tag(xml_id) or a tag(id).
It handles tags for domain engine too.

closes odoo/enterprise#50673

Task-id: 3497908
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>