[FIX] web_studio: show existing fields all the time
This commit fixes the filtering applied once a field was selected from the
studio sidebar in the search view editor. For this view, the fields must
always be draggable since they can be put in different location in the view.

Some search editor tests have been modified to this change, since more fields
are available to add in the view. An assertion has also been added to verify
that no filtering happens when dropping a field into the view.

task-3474817

closes odoo/enterprise#51243

X-original-commit: e7afe6958ea3549274c464502d0bf5e3eac71ed2
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>