[FIX] l10n_mx_edi: fix uploaded attachment rights issue in invoices
Steps to reproduce:
- Install l10n_mx_edi
- Use an user with Accountant rights in Accounting (e.g. Mitchell Admin)
- Switch to a Mexican company (e.g. ESCUELA KEMPER URGATE)
- Go to "Accounting / Customers / Invoices"
- Click on "Upload" button and upload a CFDI invoice
- Connect with another user with Accountant rights in Accounting (e.g. Marc Demo)
- Try to open the invoice created from the uploaded CFDI invoice
=> You will be able to open the invoice, but an Access Error on the attachment will be raised.

Cause:
When the CFDI invoice is uploaded, an attachment is correctly created
from it and linked to the invoice.
But another attachment is created from that document to create a
"l10n_mx_edi.document" record.
This second attachment has no reference to the invoice (i.e. its "res_model"
and "res_id" fields are not set).

opw-3579253

closes odoo/enterprise#50949

X-original-commit: 2bc01723539a075131cee24126dd4f11e2540f6a
Related: odoo/odoo#142594
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>