[IMP] currency_rate_live: add Bank of Guatemala
In Guatemala the practice is to use the exchange rates provided by the
Bank of Guatemala. They offer a webservice to fetch the daily rates that
we are adding as a Currency Rate Provider in this commit.

Further we are cleaning up the Current Rate Provider list by adding the
country code in front of the label and sorting alphabetically.

task-3581837

closes odoo/enterprise#52739

X-original-commit: 40d62fb3ce62a22e455ec98f2fe692286138786e
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>