[FIX] web_studio: fix style and traceback on field popover
Before this commit, if the name of the current model on the field
popover was to large, there was a blank space.
In addition, in debug mode, there was a traceback when trying to add
a field in the dynamic table because of props validation.

After this commit, no more blank space and the format of the SelectMenu's
items are corrected.

closes odoo/enterprise#50711

Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>