[IMP] {pos_restaurant_}appointment: make demo data realistic
Adapt slots of Table Booking appointment to match restaurant-like
hours: bookable every 30 mins, open between wednesday and sunday
included at night, and with a lunch service at noon on wednesday
and on the weekend. Also update the duration of the appointment
to 2 hours.

Add a question on dietary preferences. Call the location
to Bistr-Odoo and make it a company.

Also, for other demo data types in the appointment module,
set videoconference link to False.

Task-3593599

X-original-commit: a011a9f73eca8c1d5dfb4625432c4c35918d6f0a
Part-of: odoo/enterprise#52846