[FIX] mrp_workorder: WO state filters
The existing behavior for showing or hiding WO cards based on their
state was not really transparent to the user and had multiple bugs
in combination with the new card caching mechanism.

This PR adds a new, editable, filter in the search box to show the user
the current filter criteria for WO states and uses this opportunity to
clean some related dead code.

Part-of: odoo/enterprise#50148