[IMP] appointment: show spinner and require capacity selection
1. Add a falsy default value for capacity

If asked_capacity is not set (coming back from further step)
use a '-' option by default, not showing any time slots until
a capacity is selected.

This way, that number must be knowingly selected, preventing
booking for 1 without noticing, as this was the default value,
missing this choice.

2. Give title of questions more space

Introduced in 4f68079b535c1ea54c42ae8dabae903f88d6e87f, all the
questions have their labels inlined with the input areas. The same
is true for the 'text' question, despite having the textarea
displayed on next line. Therefore we have only a small width given
to title, despite having the whole line available. Make it use the
available space. Also, align the textarea with the rest of the form.

For the custom questions, show answers on next line to give full
line width to their title.

3. Add a spinner for time slots

As the time slots can take a while to compute and update, we add a
spinner to indicate that the slots are yet to be updated. When
capacity is enabled, '-' option is considered as missing data,
therefore it will not show the spinner as no slots should be
displayed. Only show the spinner when a day is selected, too.

Task-3593599

closes odoo/enterprise#50729

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>