[FIX] stock_barcode: scan product in multicompany
Before this commit, it was possible to scan product from a company other
than the active one.
How to reproduce:
- Have at least two active companies;
- Create a product with a barcode and set the second company in
  `company_id`;
- Now, active only the first company, go in the App Switcher, open the
  Barcode App and open an Inventory Adjustment;
- Scan the product -> The product is found and an inventory line is
  created regardless it doesn't belong to the active company.

To fix that, add the allowed companies id in the product domain in the
route `get_specific_barcode_data`. It's a backport of the following
commit: f09d7e6af1913f7c8e5d014a290cd32265ad497c
That said, the usecase is not really the same so the test is different,
even if the two fixes are really similar.

The exact same issue can happen for lot and serial numbers, a test was
also added to cover that.

opw-3582121
task-3472884

X-original-commit: 7c57ca7e9f373f1349734727a45b7d772ed4c6b1
Part-of: odoo/enterprise#53600