[FIX] stock_barcode: read product barcode as GS1
When using the GS1 nomenclature, it can happen a product's barcode is
read as a GS1 barcode for something else (for a lot or a serial number
if the barcode starts with 10 or 21 for example), which cause this
product to be unscannable if not scanned with the right prefix (the AI
for product being 01 or 02).

How to reproduce:
- In the Inventory Settings, use the Default GS1 Nomenclature;
- Create a product with an ambiguous barcode (eg.: 21000000000003, whom
  can be taken for a serial number);
- Go to the Barcode App, create a new receipt and scan the product's
  barcode -> The product is not found, nothing happens except a warning.

The same issue can happen for a package, product packaging's barcode,
lot and serial numbers aswell.

task-3472884

closes odoo/enterprise#53892

X-original-commit: b51f3b505053df760287fd891630a283aa50cc07
Part-of: #53600
Related: odoo/odoo#1147955
Related: odoo/odoo#148603
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>