[FIX] web_gantt: totals responsive for smaller precision times
Steps to reproduce:

- Install the planning app and Studio.
- Go to the planning app and trigger studio to edit.
- Change the `Day Precision` to `Quarter Hour`.
- Close and now click on `Week` and change the view to `Day`.

The issue is that since the gantt is generated based on the screen size
of the user in order to fit the whole gant in that width, when we divide
the cell in 4 or even 2, the totals are going to overlap in smaller
screens since there's no space for that division to display the totals
horizontally.

But in the other hand If we display the totals vertically and make them
have a responsive fontsize, it will be visible for pretty much every
screen size.

opw-3538039

closes odoo/enterprise#54462

X-original-commit: 0620ba1f1756a08e368a626b286afcc5eb83e223
Signed-off-by: Bastien Fafchamps (bafa) <bafa@odoo.com>
Signed-off-by: Maruan Aguerdouh Mohtar (magm) <magm@odoo.com>