[IMP] l10n_lu_reports: switch validation errors into warnings
Currently, we do a validation when an XML is being exported.
If there are several checks that fail one after another, the users
have to try to export, face the error, fix, try to export again several times.
This commit changes these into warnings.
We allow exporting XML despite these warnings, so that users are not blocked
that way.

task-3556111

closes odoo/enterprise#51500

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>