[FIX] documents: manage not access to ir.model
Steps to reproduce:
-------------------
- install documents and fleet apps;
- be a user who is not in the `base.user_admin` group;
- have admin rights to fleet and document;
- put a document in the "Fleet" workspace;
- click on "Link to a vehicle" button;

Issue:
------
- First issue:
An access error is triggered when we want to get the `model` field for an `ir.model` record.
- Second issue:
When a `Reference` field has the `model_field` option, a read on `ir.model` is triggered
and as we don't have the rights, this causes an access error.

Solution:
---------
Work on the `ir.model` record in sudo.
Do not use the `model_field` option and use a default value for the `Reference` field
which expects a string like `"model,id"`.

Note:
-----
This solution avoids adding a new access right for `ir.model` as follows:
```csv
access_ir_model,access_ir_model,base.model_ir_model,documents.group_documents_user,1,0,0,0
```

opw-3502558

closes odoo/enterprise#56147

X-original-commit: 0a1fd95643f5fc5d9782f3371bb6b452b5b01ce9
Signed-off-by: Thomas Lefebvre (thle) <thle@odoo.com>