[FIX] marketing_automation: avoid trace duplicates for new participants
When synchronizing a campaign using 'action_update_participants' traces are
created for activities considered as 'requiring sync'. However traces may
already exist for some participants that have already participated to some
activities. This leads to duplicate traces.

In this commit we now create new traces only for participants that have not
already participated to the activity to update.

Task-2981581 (MA: Fix trace duplication and various issues)
opw-2643368

X-original-commit: d408efeccdfa0e897a582ac379d96fd18db79908
Part-of: odoo/enterprise#51232
Co-authored-by: Stéphane Debauche <std@odoo.com>
Co-authored-by: Thibault Delavallée <tde@odoo.com>