[FIX] l10n_mx_edi_pos: pass order to AddInfoPopup
Prior to this commit, attempting to invoice a paid order in the POS
Restaurant module would result in an error due to 'selectedOrder' being
undefined. This issue was observed when trying to process the invoicing
of an already paid order.

This commit resolves the issue by ensuring that the order is passed as
a prop to the AddInfoPopup.

opw-3590978

closes odoo/enterprise#51726

X-original-commit: 64c5058d0f7451ef5d3ba0e74180dd4f3b12c94f
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>