[FIX] l10n_mx_edi_pos: pass order to AddInfoPopup
Prior to this commit, attempting to invoice a paid order in the POS
Restaurant module would result in an error due to 'selectedOrder' being
undefined. This issue was observed when trying to process the invoicing
of an already paid order.

This commit resolves the issue by ensuring that the order is passed as
a prop to the AddInfoPopup.

opw-3590978

closes odoo/enterprise#51433

X-original-commit: 4e0422bf53918c2c9ad87b6ea31ff0cd73563132
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>