[FIX] appointment: fix new design discrepancies
When using 'resource_time' without pictures, the name
of the variable within the resource loop in the select is
'resource', which is also the name of the resource described
in the details right column. Therefore, the last resource was
always shown. It should not be shown at all. Templates are
updated to have unique names.

Also, when selecting a user in operator screen (when website is
installed and pictures are enabled), their info is not showing on
the right column on time selection screen.

We also remove isOperator as not strcitly useful but instead
explicitely set resource to False in appointment_select_operator
in order to make sure it is not used as a selected resource in
other templates. This is necessary because of the loop on
available resources using t-as="resource". This solution avoids
renaming all instances in that template.

Task-3588251

closes odoo/enterprise#50826

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>