[FIX] whatsapp: do not raise for safe field access if unchanged
Steps to reproduce:
1. Install the Whatstapp module.
2. Login with the admin user and create a template with model contact and in the
   body add one variable.
3. now go to the variable tab and for variable select type as a field of the
   model and then select any field.
4. Make an internal user Administrator of WhatsApp.
5. Login with user created.
6. go to the created template and change the footer of the template.

This results in a Validation Error: "You are not allowed to use this field,
contact your administrator."

Reason:
When anyone writes a field of template _check_field_name is called even if
variables are not changed. If there is any field that is not in the safe fields
it raises a Validation error.

Solution:
Check WhatsApp template variable constraints for all variables of the type
field if the model is changed in the Whatsapp template. When variables are
changed then there is no need to call constraints manually from the
WhatsApp template as they are called automatically.

Also added onchange for field_type to set field_name = False if field_type
is not 'field' as it is invalid.

Task-3517290
Sub-part of Task-3608129 (Whatsapp: Moultifix !)

X-original-commit: b9dba97271c4670d2831adad821459710efcb357
Part-of: odoo/enterprise#52840