[FIX] account_reports: fix audit of analytic columns in report
When grouping by analytic plan and/or account in a report, clicking the amount always opened an empty tree view, except for the default "Projects" plan which properly displayed the analytic lines making the amount.

This was due to the fact the domain was built using field account_id of account.analytic.line, which is only used for the default plan (other plans dynamically create fields on account.analytic.line to store their chosen account). auto_account_id is overridden in the model to allow searching on all plans, and should hence be used in this context.

closes odoo/enterprise#55752

X-original-commit: 2c554a075c77e990755d8177421fa2ff39402d35
Signed-off-by: William André (wan) <wan@odoo.com>