[FIX] mrp_workorder: quick register production only when not set
When some value is set on a register production line, and we try to use
the quick action, it will override the value we have currently set.
As this is confusing for the user, we now open the popup instead when
trying to use the quick action when some value is already set.

Task: 3549953 (nr 9)
Part-of: odoo/enterprise#52456