[IMP] sign: replace jSignature library
The jSignature library depends on jQuery. As we plan to remove jQuery in
the future, jSignature is replaced by another (simpler) library,
with fewer dependencies.

Enters signature_pad [1], which looks well maintained, has no external
dependency and is (roughly) 1/3 of the size.

This commit adapts the related components.

Note: removed the comment in manifest related to the jQuery's usage, as
it is stil used anyway.

task-3599215

[1]: https://github.com/szimek/signature_pad

closes odoo/enterprise#50901

Related: odoo/odoo#142507
Signed-off-by: Romain Estievenart (res) <res@odoo.com>