[IMP] account_reports, l10n_be_reports: remove tax_report_control_error dead code
The field tax_report_control_error defined on account.move was previously used to display a warning message on the tax closing's form view when some of the controls defined on the report failed. This was only used in l10n_be_reports at this point.

From 16.4 on, this feature did not work anymore, because of the introduction of sections, and the fact the options dict is generated separately from the lines, and hence cannot be modified while these are being generated.

https://github.com/odoo/enterprise/commit/9b3bb03f27a142d367310e6def856efd194b3006 already removed such a warning on the Belgian XML export wizard, and introduced a warning banner directly on the report to replace it. With this commit, we make the choice to remove the wanring on the closing move as well ; any regular tax closing flow should anyway go through the report, and then view the warning displayed there.

Another PR will follow in master to finish removing the unneeded fields.

closes odoo/enterprise#51235

X-original-commit: db4b9a10e2fa7c8dff1f21a667e1b702e826d347
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: John Laterre (jol) <jol@odoo.com>