[FIX] social: prevent the read time out error in compute statistics
While computing statistics with the function(_compute_statistics) in
the social marketing module, which is facing connectivity issues due
to a loss of network or slow internet, the page is getting read time out.
This is happening when the page is being refreshed.

This commit avoids showing an error to the end user; instead, it shows an error
in the log as a warning and ignore `Timeout and ConnectionError` globally,
which are generated from social media.

Apart from this, apply the same changes to the previously ignored ReadTimeout
error in social with  https://github.com/odoo/enterprise/pull/51359.

sentry-4649545565

closes odoo/enterprise#52484

X-original-commit: 8df20eae1ae94141b7d4cf0b8aea34865b757bf8
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>