[FIX] sale_planning: remove step to click on plan button in gantt
Before this commit, a recent generic change removes the "add" and
"plan" buttons in each gantt cell to avoid having too many elements
available in the gantt view. That changes has not been adapted for the
sale_planning tour and so the step failed because it no longer found
plan button.

This commit removes the step failed and adapt the previous step to
just click on the grid cell to display the plan dialog.

runbot-25881

closes odoo/enterprise#51053

X-original-commit: 432a8b071d5bef1fed3107b468cbbc59f408d878
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>