[FIX] planning: freeze time for test_shift_creation_from_role
Before this commit, the `test_shift_creation_from_role` test failed
during the weekend because the working schedule of the current company
considers the weekend is not a working day and so instead of creating a
planning template during 2 days, it will create a planning template during
one day.

This commit freezes the time to be sure, the test is executed during a
working day to stay valid and keep the same behavior to test.

runobt-41977

closes odoo/enterprise#51075

X-original-commit: 23ba24ae77ca405f74e9188cd7a96eabaacdc334
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>