[IMP] approvals: create rules for admin & unlink
Make basic rights for administrator to gives them permissions
to delete approval requests before it was limited to the domain for
officer (i.e. only your own approval).
The products was also preventing the deletion of the approvals thus
the unlink method had to be rewritten to make it possible.

task-3463942

closes odoo/enterprise#53480

X-original-commit: 2b9046ce13a50617f6ddcd68b84ae7d55190e5af
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Florent Maisse (mafl) <mafl@odoo.com>