[FIX] web_studio: display the correct fields for fold_field option
This commit fixes the display of available fields for the fold_field option
in the view editor, when editing the statusbar widget. Now, the model from the
corresponding relation is used instead of the model used directly in the
current view. An async call is executed specifically for this option, requiring
the state to be set only when the result is known.

This fix highlighted another bug, since subOptions have been introduced by
commit (1). A copy of the options was set in the subOptions array. Since the
compute of field choices is not instant, the option would appear in the
sidebar without the selection menu. The asynchronous compute made things
even worse. Now, the complete info concerning an option is directly set under
'attributesForCurrentTypeAndWidget', and the subOptions are retrieved from this
object, having the latest keys available for the rendering (e.g. 'choices').

To do this, it was needed to create a deep copy of the options descriptions, since
we don't want to alter directly those objects.

A test has been added to check that the choices are displayed properly, even
when using suboptions.

task-3599722

1) 47e8af201143ebe9a0d452ac481e44b546d74150

closes odoo/enterprise#53813

X-original-commit: ffb5aa25eb76c2229fa3034b67ff5cfdf8221316
Related: odoo/odoo#148481
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>