[FIX] whatsapp: search for related last messages based on associated account
Steps to reproduce:
1. Install the WhatsApp module.
2. Create two WhatsApp accounts.
3. Send a template message from the first account to a recipient's number.
4. Reply to the message from the WhatsApp application.
5. A new channel will be created for this conversation linked to the first
   account.
6. Using the WhatsApp application on the same number, send a message to the
   WhatsApp number of the second account.

In Odoo, a new channel will be created with the related message set to the last
template message of the first conversation and linked to the second account.

Reason:
This issue arises because the system searches for the related last message
without considering the associated WhatsApp account.

Solution:
To address this problem, update the search logic to consider the associated
WhatsApp account when searching for related last messages and correctly
associate it with the new channel.

Task-3520468
Sub-part of Task-3608129 (Whatsapp: Moultifix !)

X-original-commit: 05f4c368360d79b5afe1d403a065f725518b717d
Part-of: odoo/enterprise#52840